+++ src/main/java/chess/domain/Board.java
@@ -1,4 +1,4 @@
-package chess.board;
+package chess.domain;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -7,7 +7,10 @@ import java.util.Map;
 
 import com.google.common.collect.Lists;
 
-import chess.piece.Piece;
+import chess.domain.position.File;
+import chess.domain.position.Rank;
+import chess.domain.position.Square;
+import chess.domain.piece.Piece;
 
 public class Board {
     private final Map<Square, Piece> board;
