+++ src/main/java/chess/domain/piece/Color.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.domain.piece;
 
 public enum Color {
     BLACK,
