+++ src/main/java/chess/domain/piece/None.java
@@ -1,6 +1,6 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.board.Square;
+import chess.domain.position.Square;
 
 public final class None extends Piece {
     private static final String NONE = "·";
