+++ src/main/java/chess/domain/position/Square.java
@@ -1,4 +1,4 @@
-package chess.board;
+package chess.domain.position;
 
 import java.util.Objects;
