+++ src/test/java/chess/domain/board/FileTest.java
@@ -1,9 +1,11 @@
-package chess.board;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import chess.domain.position.File;
 
 public class FileTest {
     @Test
