+++ src/test/java/chess/domain/board/RankTest.java
@@ -1,9 +1,11 @@
-package chess.board;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import chess.domain.position.Rank;
 
 public class RankTest {
     @Test
