+++ src/test/java/chess/domain/board/SquareTest.java
@@ -1,9 +1,13 @@
-package chess.board;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import chess.domain.position.File;
+import chess.domain.position.Rank;
+import chess.domain.position.Square;
 
 public class SquareTest {
     @Test
